add syncing content to gh action

+++ sync_content.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
-rm -rf content_temp
-rm -rf src/content/*
-git clone git@github.com:pompeybug/content.git content_temp
+if [ "$1" != '--pages' ];
+then
+  rm -rf content_temp
+  rm -rf src/content/*/
+  git clone git@github.com:pompeybug/content.git content_temp
+fi
 mv content_temp/* src/content
 rm -rf content_temp
